Cache last calculated number of available items
This change backs out the previous consumer-side caching, and replaces
it with a less invasive caching mechanism that allows the publisher task
to also avoid retrieving sequence values from dependent consumers when
they already established availability of enough elements during a
previous wait.

Additionally, the newly cached number of available slots is used to
reduce the number of times that the consumer publishes an updated
sequence number to downstream consumers. In hindsight, this could have
been done in a separate commit from the availability caching.

Combined, these changes more than double throughput compared to the last
version of the code.