Add a wait strategy that sleeps indefinitely while waiting
In cases where conserving CPU time is a priority, it's preferable to
have a wait strategy that eventually goes to sleep until the next item
is released. This commit adds BlockingWaitStrategy, which does that.

This is similar to the BlockingWaitStrategy from the LMAX disruptor, but
with a double-checked algorithm that leads to far better performance
during bursts of events (short or sustained).