Move ring buffer reference into SinglePublisherSequenceBarrier
It is fairly clear to me now that it was misguided to try to separate
the barrier types from the ring buffer itself. Although currently, the only
necessary parameter is the buffer size, a mutable reference to the ring buffer itself is required to implement resizing. With that case, it is simpler to just store the ring buffer as a member of the barrier type, not as a member of the Publisher and Consumer types.