RingBufferTrait: Add Send as a type bound for T
Types that implement RingBufferTrait can be used to pass elements of T
between tasks, which would be incorrect for non-Send instances of T.
This change fixes the oversight in RingBufferTrait and callers. Note
that since the public API was already enforcing this bound on T, it was
not possible to take advantage of this loophole.