Factor out a RingBufferOps trait from RingBufferTrait
The sole difference is that RingBufferOps doesn't inherit from Clone.
This allows the SequenceBarrier trait to be able to wrap these
operations, without forcing implementors to also implement clone.  It
would make just as much sense for SequenceBarrier to be cloneable, but
until there's a use case for it, it's probably better to avoid adding
dead code.