Replace SequenceBarrier::get_ring_buffer with RingBufferOps supertrait
This change removes the get_ring_buffer fn from SequenceBarrier, and
replaces it by inheriting from RingBufferOps instead. This allows
implementors of SequenceBarrier to easily wrap around the ring buffer
operations if necessary (such as for resizing), and reduces the number
of type parameters that callers have to pass to SequenceBarrier.