RingBufferData.take: Improve error message if value is null
Previously, if a misbehaving caller were to call take() before a value
has been set in a slot, the task would fail, but the error message
didn't include enough information to help diagnose the issue. This
change customizes the error message to print out the sequence number of
the slot that was being retrieved.