Quick and dirty port to latest rustc
This change ports the code from rust commit
1bfa45cfd9d7c8b2dcc0ea7a42396624e8b5a31b (2013-11-12) to commit
25431774a933a3b0c5212ea45150660544dec1ec (today).

This commit disables the benchmark for SpinWaitStrategy until I find
time to rewrite the benchmark to use 1:1 task scheduling for it again.
The YieldWaitStrategy benchmark produces results that are almost as high
as SpinWaitStrategy's, so this is acceptable in the short term.