Fix borrowing violation that the compiler previously allowed
The compiler previously allowed programs to use destructuring lets to
borrow a mutable reference to the field(s) inside a tuple struct, and
then borrow a reference to the tuple struct itself while the first
borrow was still live. This change fixes one of the tests, which was
previously doing this for convenience.