Factor out traits for the disruptor module's public interface
Previously, the public API has exposed a SequenceBarrier generic type
parameter, even thought that isn't very relevant to callers. This change
prepares to fix this by factoring out three traits that define the core
of the public API: Publisher, Consumer, and FinalConsumer.

Unfortunately, as a result of this change, callers now have to import
the traits to be able to use the API.