Remove SequenceBarrier trait from public interface of disruptor module
This change renames the existing SinglePublisher, SingleConsumer, and
SingleFinalConsumer types into GenericPublisher, GenericConsumer, and
GenericFinalConsumer, and adds new types with the same name that do not
expose SequenceBarrier as a type parameter, and do not support dynamic
resizing of the ring buffer. The resizing support is exposed in three
new types, which are also monomorphic with respect to SequenceBarrier:
SingleResizingPublisher, SingleResizingConsumer, and
SingleResizingFinalConsumer.

Although this change makes the code more complicated in the short term,
I hope that after rearranging the internals to line up more closely with
the public interface, things will be simpler than they were before
adding all this complication.