auto-rebuild: limit wasted cycles on unicast_throughput_benchmark
This change hardcodes a smaller number of iterations on
unicast_throughput_benchmark. In theory, it would make sense to expose
the number of iterations as a parameter to this script, but if
unicast_throughput_benchmark is ported to a benchmark (using Criterion's
iter_custom timing loop), it'll eliminate the need to manually choose a
number of iterations, so it's not worth the effort building any
infrastructure around that.